// types/release.ts
export type AppLatestServer = {
  id: string
  type: 'app'
  mainRelease: AppLatestJSONLatest
  lastUpdated: number
  totalDownloads: number
  pastReleases?: PastReleaseInfo[]
}

/**
 * The outdated version of the release file used for legacy migrations
 * @deprecated - use 0.11.8
 */
export type AppReleaseFile0108 = {
  /**
   * @deprecated - use 0.11.11 (latest)
   */
  version: '0.10.0'
  repositories: string[]
  references: AppReleaseCommunity[]
  releases: AppReleaseMeta[]
  timestamp: number
}

export type PastReleaseInfo = {
  tag: string
  downloads: number
  size: number
  name: string
  download_url: string
  created_at: string
}